+++ App/auth/email.py
@@ -1,0 +1,15 @@
+from flask import render_template, current_app
+from flask_babel import _
+from App.email import send_email
+
+
+def send_password_reset_email(user):
+    token = user.get_reset_password_token()
+    """Sends an email to reset a user's password"""
+    send_email(_('[Blog] Reset Your Password'),
+               sender=current_app.config['ADMINS'][0],
+               recipients=[user.email],
+               text_body=render_template('email/reset_password.txt',
+                                         user=user, token=token),
+               html_body=render_template('email/reset_password.html',
+                                         user=user, token=token))
