+++ App/auth/forms.py
@@ -39,31 +39,6 @@ class SignUpForm(FlaskForm):
             raise ValidationError(_('Email already exists!'))
 
 
-class EditProfileForm(FlaskForm):
-    """Enables a users profile to be set and edited"""
-    username = StringField(_l('Username'), validators=[DataRequired()])
-    about_me = TextAreaField(_l('About me'), validators=[
-                             Length(min=0, max=140)])
-    submit = SubmitField(_l('Submit'))
-
-    def __init__(self, original_username, *args, **kwargs):
-        super(EditProfileForm, self).__init__(*args, **kwargs)
-        self.original_username = original_username
-
-     # prevent duplicate usernames
-    def validate_username(self, username):
-        if username.data != self.original_username:
-            user = User.query.filter_by(username=username.data).first()
-            if user is not None:
-                raise ValidationError(_('Please use a different username'))
-
-
-class PostForm(FlaskForm):
-    post = TextAreaField(_l("What's on your mind?"), validators=[
-        DataRequired(), Length(min=1, max=140)])
-    submit = SubmitField(_l('Submit'))
-
-
 class ResetPasswordRequestForm(FlaskForm):
     email = StringField(_l('Email'), validators=[DataRequired(), Email()])
     submit = SubmitField(_l('Request Password Reset'))
