+++ App/cli.py
@@ -1,9 +1,12 @@
 import os
 import click
-from App import app
+from flask import Blueprint
 
 
-@app.cli.group()
+bp = Blueprint("cli", __name__, cli_group=None)
+
+
+@bp.cli.group()
 def translate():
     """Translation and localization commands."""
     pass
