+++ App/models.py
@@ -1,9 +1,9 @@
 from datetime import datetime
 from hashlib import md5
+from flask import current_app
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import UserMixin
 from App import db, login
-from App import app
 import jwt  # type: ignore
 from time import time
 
@@ -80,12 +80,12 @@ class User(UserMixin, db.Model):
         return jwt.encode(
             # payload contains user id and current time + 10min
             {'reset_password': self.id, 'exp': time() + expires_in},
-            app.config['SECRET_KEY'], algorithm='HS256')
+            current_app.config['SECRET_KEY'], algorithm='HS256')
 
     @staticmethod
     def verify_reset_password_token(token):
         try:
-            id = jwt.decode(token, app.config['SECRET_KEY'],
+            id = jwt.decode(token, current_app.config['SECRET_KEY'],
                             algorithms=['HS256'])['reset_password']
         except:
             return
@@ -108,7 +108,6 @@ class Post(db.Model):
     timestamp = db.Column(db.DateTime, index=True,
                           default=datetime.now)  # refer back utcnow deprecated
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
-    language = db.Column(db.String(5))
 
     def __repr__(self):
         return f'<post {self.body}'
