+++ App/templates/auth/login.html
@@ -4,9 +4,9 @@
 {% block content %}
 <h1>{{ _('Sign In') }}</h1>
 {{ wtf.quick_form(form) }}
-<p>{{ _('New User?') }} <a href="{{ url_for('SignUp') }}">{{ _('Click to Register!') }}</a></p>
+<p>{{ _('New User?') }} <a href="{{ url_for('auth.SignUp') }}">{{ _('Click to Register!') }}</a></p>
 <p>
   {{ _('Forgot Your Password?') }}
-  <a href="{{ url_for('reset_password_request') }}">{{ _('Click to Reset It') }}</a>
+  <a href="{{ url_for('auth.reset_password_request') }}">{{ _('Click to Reset It') }}</a>
 </p>
 {% endblock %}
