+++ App/templates/auth/reset_password_request.html
