+++ App/templates/errors/500.html
@@ -3,5 +3,5 @@
 {% block content %}
 <h1>{{ _('An expected Error has occurred') }}</h1>
 <p>{{ _('The adminstrator has been notified. Sorry for the inconvinience!') }}</p>
-<p><a href="{{ url_for('index') }}">{{ _('Back') }}</a></p>
+<p><a href="{{ url_for('main.index') }}">{{ _('Back') }}</a></p>
 {% endblock %}
