+++ requirements.txt
@@ -16,3 +16,4 @@ Werkzeug==3.0.4
 WTForms==3.1.2
 Babel==2.16.0
 flask_babel==4.0.0
+python-dotenv==1.0.1
